requirements_dev.txt
bumpversion==0.5.3
wheel==0.24.0
djangorestframework==3.3.2
funcy==1.7.1
